+++ Runtime/UI/Components/AssetTargetField.cs
@@ -2,6 +2,9 @@ using System;
 using Unity.AppUI.Core;
 using UnityEngine;
 using UnityEngine.UIElements;
+#if ENABLE_RUNTIME_DATA_BINDINGS
+using Unity.Properties;
+#endif
 
 namespace Unity.AppUI.UI
 {
@@ -15,6 +18,8 @@ namespace Unity.AppUI.UI
     partial class AssetTargetField : BaseVisualElement, IInputElement<AssetReference>, ISizeableElement, IPressable
     {
 #if ENABLE_RUNTIME_DATA_BINDINGS
+        internal static readonly BindingId typeProperty = nameof(type);
+
         internal static readonly BindingId sizeProperty = nameof(size);
 
         internal static readonly BindingId valueProperty = nameof(value);
@@ -112,11 +117,15 @@ namespace Unity.AppUI.UI
             SetValueWithoutNotify(null);
         }
 
+        /// <inheritdoc/>
         public override VisualElement contentContainer => null;
 
         /// <summary>
         /// Clickable Manipulator for this AssetTargetField.
         /// </summary>
+#if ENABLE_RUNTIME_DATA_BINDINGS
+        [CreateProperty]
+#endif
         public Pressable clickable
         {
             get => m_Clickable;
@@ -136,11 +145,19 @@ namespace Unity.AppUI.UI
             }
         }
 
+        /// <summary>
+        /// The type of the AssetReference that this field accepts.
+        /// This is used to filter the assets that can be assigned to this field.
+        /// </summary>
+#if ENABLE_RUNTIME_DATA_BINDINGS
+        [CreateProperty]
+#endif
         public Type type
         {
             get => m_Type;
             set
             {
+                var changed = m_Type != value;
                 m_Type = value;
                 if (m_AssetReference != null && m_Type != null)
                 {
@@ -151,26 +168,70 @@ namespace Unity.AppUI.UI
 
                 m_IconElement.iconName = m_Type?.Name.ToLower();
                 m_TypeLabelElement.text = m_Type?.Name.ToUpper();
+
+#if  ENABLE_RUNTIME_DATA_BINDINGS
+                if (changed)
+                    NotifyPropertyChanged(in typeProperty);
+#endif
             }
         }
 
+        /// <summary>
+        /// The size of the element.
+        /// </summary>
+#if ENABLE_RUNTIME_DATA_BINDINGS
+        [CreateProperty]
+#endif
+#if ENABLE_UXML_SERIALIZED_DATA
+        [UxmlAttribute]
+#endif
         public Size size
         {
             get => m_Size;
             set
             {
+                var changed = m_Size != value;
                 RemoveFromClassList(GetSizeUssClassName(m_Size));
                 m_Size = value;
                 AddToClassList(GetSizeUssClassName(m_Size));
+#if ENABLE_RUNTIME_DATA_BINDINGS
+                if (changed)
+                    NotifyPropertyChanged(in sizeProperty);
+#endif
             }
         }
 
+        /// <summary>
+        /// Whether the current value of the AssetTargetField is valid or not.
+        /// This is determined by the validateValue function.
+        /// </summary>
+#if ENABLE_RUNTIME_DATA_BINDINGS
+        [CreateProperty]
+#endif
+#if ENABLE_UXML_SERIALIZED_DATA
+        [UxmlAttribute]
+#endif
         public bool invalid
         {
             get => ClassListContains(Styles.invalidUssClassName);
-            set => EnableInClassList(Styles.invalidUssClassName, value);
+            set
+            {
+                var changed = invalid != value;
+                EnableInClassList(Styles.invalidUssClassName, value);
+#if ENABLE_RUNTIME_DATA_BINDINGS
+                if (changed)
+                    NotifyPropertyChanged(in invalidProperty);
+#endif
+            }
         }
 
+        /// <summary>
+        /// A function that validates the value of the AssetTargetField.
+        /// It returns true if the value is valid, false otherwise.
+        /// </summary>
+#if ENABLE_RUNTIME_DATA_BINDINGS
+        [CreateProperty]
+#endif
         public Func<AssetReference, bool> validateValue
         {
             get => m_ValidateValue;
@@ -187,6 +248,10 @@ namespace Unity.AppUI.UI
             }
         }
 
+        /// <summary>
+        /// Sets the value of the AssetTargetField without sending a change event.
+        /// </summary>
+        /// <param name="newValue"> The new value to set.</param>
         public void SetValueWithoutNotify(AssetReference newValue)
         {
             m_AssetReference = newValue;
@@ -194,6 +259,12 @@ namespace Unity.AppUI.UI
             if (validateValue != null) invalid = !validateValue(m_AssetReference);
         }
 
+        /// <summary>
+        /// The current value of the AssetTargetField.
+        /// </summary>
+#if ENABLE_RUNTIME_DATA_BINDINGS
+        [CreateProperty]
+#endif
         public AssetReference value
         {
             get => m_AssetReference;
@@ -205,10 +276,16 @@ namespace Unity.AppUI.UI
                 evt.target = this;
                 SetValueWithoutNotify(value);
                 SendEvent(evt);
+#if ENABLE_RUNTIME_DATA_BINDINGS
+                NotifyPropertyChanged(in valueProperty);
+#endif
             }
         }
 
 #if ENABLE_UXML_TRAITS
+        /// <summary>
+        /// The UXML factory for the <see cref="AssetTargetField"/>.
+        /// </summary>
         public new class UxmlFactory : UxmlFactory<AssetTargetField, UxmlTraits> { }
 
         /// <summary>