+++ Runtime/UI/Components/NavigationRailItem.cs
@@ -123,6 +123,9 @@ namespace Unity.AppUI.UI
         /// <summary>
         /// Clickable Manipulator for this NavigationRailItem.
         /// </summary>
+#if ENABLE_RUNTIME_DATA_BINDINGS
+        [CreateProperty]
+#endif
         public Pressable clickable
         {
             get => m_Clickable;