+++ Runtime/UI/Components/Toggle.cs
@@ -117,6 +117,9 @@ namespace Unity.AppUI.UI
         /// <summary>
         /// Clickable Manipulator for this Toggle.
         /// </summary>
+#if ENABLE_RUNTIME_DATA_BINDINGS
+        [CreateProperty]
+#endif
         public Pressable clickable
         {
             get => m_Clickable;